added delete User Route

+++ controllers/userController.js
@@ -43,10 +43,15 @@ module.exports = {
    },
     // Delete User by ID 
     deleteUser(req, res) {
-    User.findByIdAndDelete(
-        { _id: req.params.userId }
-    )
-   }
+    User.findByIdAndDelete({ _id: req.params.userId })
+    .then((user) => 
+        !user
+            ? res.status(404).json({ message: 'No User found withthis ID!'})
+            : Thought.deleteMany({_id: {$in: user.thought} })
+            )
+            .then(() => res.json({ message: 'User and assosiated thought was deleted!'}))
+            .catch((err) => res.status(500).json(err));
+   },
 
 };
